move node reference to application directory
