+++ src/OpcUaStackCore/Application/NodeReference.h
